moved all necessary files to the shared directory
