+++ public_html/api/json-snippets/sign-in.json
@@ -1,3 +1,3 @@
 [
-  {"profileEmail":"gkephart@cnm.edu","profilePassword":"password1"}
+  {"profileEmail":"gkephart@cnm.edu","profilePassword":"password"}
 ]
