+++ public_html/templates/list-tweets.html
@@ -11,9 +11,9 @@
 			</div>
 		</div>
 		<div class="col-lg-4">
-			<div *ngIf="createTweetForm.controls.tweetContent?.invalid && createTweetForm.controls.tweetContent?.touched" class="alert alert-danger">
+			<!-- <div *ngIf="createTweetForm.controls.tweetContent?.invalid && createTweetForm.controls.tweetContent?.touched" class="alert alert-danger">
 				<p *ngIf="createTweetForm.controls.tweetContent?.errors.maxlength">Tweet is way to long</p>
-			</div>
+			</div>-->
 		</div>
 
 		<div class=" col-lg-8">
