+++ src/app/components/signin.component.ts
@@ -6,9 +6,9 @@ import {Component, ViewChild, EventEmitter, Output} from "@angular/core";
 
 import {Router} from "@angular/router";
 import {Observable} from "rxjs/Observable"
-import {Status} from "../classes/status";
+import {Status} from "../shared/classes/status";
 import {SignInService} from "../services/sign.in.service";
-import {SignIn} from "../classes/sign.in";
+import {SignIn} from "../shared/classes/sign.in";
 import {CookieService} from "ng2-cookies";
 declare var $: any;
