+++ src/app/shared/classes/profile.ts
