+++ src/app/shared/classes/status.ts
