+++ src/app/shared/classes/tweet.ts
